feat: Use refs for add and delete
This updates useLogs to provide references to the log documents and maintain a reference to the log collection. With those, we can supply the document ref directly to `deleteLog` and capture the logs reference directly in `addLog`.
This means we don't have to pass the firebase context around.

I also set up the delete buttons so that they will show a spinner while the delete is in progress, but even with network throttling, delete seems to finish instantaneously. So useInProgress hasn't really been tested yet.