Add react firebase hooks (#41)
* chore: add react-firebase-hooks dependency

* refactor: Make a hooks folder for custom hooks

* feat: Use useAuthState in NavBar

* feat: App uses useAuthState and shows Loading

* chore: Import SVG spinners

* feat: Add spinner animation to Loading text

* refactor: useLogin is now unused

* feat: Use `useSignInWithEmailAndPassword` on `LoginForm`

Also take advantage of the `loading` state to show a spinner on the Log In button, and disable the button if the login request is in flight.

* refactor: Reuse `Spinner` in `App`

* fix: Don't afford click if buttons are disabled

* feat: Use react firebase hooks in `SignupForm`

* refactor: `lastError` is no longer needed

* fix: Spinner doesn't respect light/dark mode

* refactor: Extract useAccount hook from logs-collection

* docs: Add ADR for react-firebase-hooks

* fix: Can't see text boxes in light mode

* docs: Add `reactfire` to the ADR now that I know about it

* refactor: Extract draft features.

Draft features are account level

* refactor: logs-collection is a useLogs hook

* refactor: use-account instead of useAccount

* feat: Use refs for add and delete

This updates useLogs to provide references to the log documents and maintain a reference to the log collection. With those, we can supply the document ref directly to `deleteLog` and capture the logs reference directly in `addLog`.
This means we don't have to pass the firebase context around.

I also set up the delete buttons so that they will show a spinner while the delete is in progress, but even with network throttling, delete seems to finish instantaneously. So useInProgress hasn't really been tested yet.

* feat: Block the Add Entry button while adding

* feat: useAccount uses react-firebase-hooks

The only weird one left is `saveMruListAndDeleteDraft`. It's probably an indication that the segmentation between accounts and logs is actually artificial.

* feat: Create a data converter for logs

This lets the hooks expose the data as my custom ILog type.