test: Switch to one large e2e test
This is the easiest way to deal with intermittency introduced by reloading the page rapidly in between test cases.
Ultimately, this series of steps was already tighly coupled. In the future, tests ought to be decoupled, but this test is in place to eliminate the need to manually test the basic scenario.