test: Add cypress tests for the happy path. (Fixes #42) (#43)
* feat: Add cypress

* feat: Generated a spec file from the UI

* feat: Add cypress tests to github actions

* fix: Use the correct start command

* WIP: try runing firebase emulators.

* WIP: Try combining firebase-tools install and ts build

* WIP: Try installing firebase tools separately

* feat: Configure cypress to run a dev server, emulators, and then tests

* docs: Add info on how to run and test

* test: Update the test to open the page

* Add firebase auth emulator

Modified the list of emulators using `firebase init emulators` to select just firestore and auth.

* docs: Note how to install firebase tools

* fix: The emulators can refuse connection if you use localhost instead of 127.0.0.1

* test: The happy path test logs in

* test: Add TS types in cypress with a config

* test: Add a test that logs out and logs back in

* test: Add create & submit draft tests

These are currently broken because firestore isn't getting updated by the tests!
I also

* test: Delete the user that was created after all the tests run

It's annoying to have to do this manually after each test run. Eventually, there will be UI to do this.

* test: Clear the firestore DB when tests are complete.

* fix: Turn on the feature that lets cypress connect to the firestore emulator.

More explanation comes from this stack overflow where I added an update based on firebase v9: https://stackoverflow.com/a/75452030/7783436

* test: Complete happy path including deleting a note

* test: Use the same commands in PRs as `npm run cypress:auto`

* test: Try having cypress GHA runner wait on the emulators and dev server

* test: Try waiting on the emulator UI instead

* refactor: Remove unncessary script and image

* test: Reduce intermittency in delete

Delete may not be working because the submit note test doesn't give the firebase library a chance to persist the note. Try a 500ms wait to reduce the intermittency.

If this works, cypress should check that the note shows up in the emulator, rather than waiting.

* test: Target the hosting emulator instead

This makes it possible to build the app with `--mode emulators`. It uses the firebase hosting emulator instead of the vite dev server as the target for cypress.

* test: typo in build script

* fix: Logging out after signing up returns to signup, not login

* test: Switch to one large e2e test

This is the easiest way to deal with intermittency introduced by reloading the page rapidly in between test cases.
Ultimately, this series of steps was already tighly coupled. In the future, tests ought to be decoupled, but this test is in place to eliminate the need to manually test the basic scenario.

* test: Wait an extra 400ms for draft save

* test: Go back to multiple its

The tests are failing in GHA now. I think it started when I removed the multiple test cases. Try switching back to those.

Revert "test: Switch to one large e2e test"

This reverts commit 6fc14042bbd5eb9016b3d8b4fea925f0a48d94cf.

* docs: Add some notes on cypress and dev server details

Also, give a brief summary of how implementing cypress went.

* docs: Exclude markdown changes and update cypress history doc

* Update docs/history-cypress.md

* test: Use projectId from config for cleaning the DB

* chore: gitignore service accounts