feat: Hide log table by default
Important for testing. This forces the time entry to rely on the list cache for its current list state. When the log table is hidden by default, there isn't anything on the page that loads the full collection of logs, and so if the list cache is empty, nothing will fill it.

The next step is to detect this case automatically and force-update the cache (with the hope that we only have to query all the logs once, to prime the cache).