feat(lists): Cache lists (#59)
* WIP

* fix: Reuse account ref for the list cache

* fix: useLogs not required for lists now.

- Time entry form uses the list cache from Account.
- Factor addLog out of useLogs

* feat: Hide log table by default

Important for testing. This forces the time entry to rely on the list cache for its current list state. When the log table is hidden by default, there isn't anything on the page that loads the full collection of logs, and so if the list cache is empty, nothing will fill it.

The next step is to detect this case automatically and force-update the cache (with the hope that we only have to query all the logs once, to prime the cache).

* refactor: Extract logsCollection from hooks into data.

* feat: Self-healing list cache

- When the list cache is absent, we'll load all the docs once and heal it.
- If a list is missing from the cache and we load logs and discover it, we'll add it.
- Whenever we add a log, we'll make sure the list is in the cache.

* feat: Use the list cache for filtering.

Now we only use the lists from the logs to ensure that the account list cache is up-to-date.

* fix: Populate with the default list

If the account has been created but there aren't any logs, add the default list to ensure it's there.

* test: Update tests to show the logs before trying to view them.

Also remove the unnecessary, failing loading spinner test.

* tests: Try upgrading cypress to clear CI-only test failures

* test: Get GHA debug logs

* test: Can't ignore workflows files if we want to debug them!

* test: Debug logging wasn't enough; record to cypress's cloud service to see what's wrong.

* test: Ignore mp4 files

* test: Try waiting slightly longer for autosave.